requirements.txt
discord.py==1.7.3
quart==0.16.2
